Add files via upload
This project is a simple interactive bubble-click game created using HTML, CSS, and JavaScript. The goal of the game is to click on the correct number displayed as the "Hit" value before the timer runs out. The player aims to achieve the highest possible score within 60 seconds by quickly and accurately clicking the correct bubbles.